Refactor HistoricalRecord
* Refactor HistoricalRecord to FantasyLeagues.HistorialRecord
* See #797
