+++ test/ex338/fantasy_leagues_test.exs
@@ -1,6 +1,6 @@
 defmodule Ex338.FantasyLeaguesTest do
   use Ex338.DataCase
-  alias Ex338.FantasyLeagues
+  alias Ex338.{FantasyLeagues}
 
   test "create_future_picks/2 create future picks for teams" do
     league = insert(:fantasy_league)
@@ -87,6 +87,34 @@ defmodule Ex338.FantasyLeaguesTest do
       assert team_1_result.points == team_1_points
       assert team_1_result.winnings == team_1_winnings
       assert result_c.id == league3.id
+    end
+  end
+
+  describe "get_current_all_time_records/0" do
+    test "only returns current all time records" do
+      _all_time3 = insert(:historical_record, type: "all_time", archived: false, order: 3.0)
+      _all_time1 = insert(:historical_record, type: "all_time", archived: false, order: 1.0)
+      _all_time2 = insert(:historical_record, type: "all_time", archived: false, order: 2.0)
+      insert(:historical_record, type: "season", order: 4.0)
+      insert(:historical_record, type: "all_time", archived: true, order: 5.0)
+
+      results = FantasyLeagues.get_current_all_time_records()
+
+      assert Enum.map(results, & &1.order) == [1.0, 2.0, 3.0]
+    end
+  end
+
+  describe "get_current_season_records/0" do
+    test "only returns current single season records" do
+      _season1 = insert(:historical_record, type: "season", archived: false, order: 3.0)
+      _season2 = insert(:historical_record, type: "season", archived: false, order: 1.0)
+      _season3 = insert(:historical_record, type: "season", archived: false, order: 2.0)
+      insert(:historical_record, type: "all_time", order: 4.0)
+      insert(:historical_record, type: "season", archived: true, order: 5.0)
+
+      results = FantasyLeagues.get_current_season_records()
+
+      assert Enum.map(results, & &1.order) == [1.0, 2.0, 3.0]
     end
   end
